requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==25.1.0
cachetools==5.3.3
chardet==5.1.0
charset-normalizer==3.4.1
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dm-tree==0.1.9
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.5.4
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.13.0
humanfriendly==10.0
idna==3.6
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.3.2
keras==3.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.14.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
patchify==0.2.3
pillow==10.2.0
progressbar2==4.5.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.19.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
scikit-build==0.18.1
scikit-image==0.25.2
scikit-learn==1.4.1.post1
scipy==1.12.0
SimpleITK==2.3.1
six==1.16.0
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
tensorlayer==2.2.5
tensorrt==10.13.0.35
tensorrt-cu12==10.13.0.35
tensorrt_cu12_bindings==10.13.0.35
tensorrt_cu12_libs==10.13.0.35
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2025.3.13
tomli==2.2.1
tqdm==4.67.1
triton==3.2.0
typeguard==2.13.3
typing_extensions==4.10.0
tzdata==2025.2
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
